Refactor into note controller
